requirements.txt
click>=8.1.7
jinja2>=3.1.3
pyyaml>=6.0.1